% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephrase_text.R
\name{rephrase_text}
\alias{rephrase_text}
\title{ChatGPT: Rephrase text}
\usage{
rephrase_text(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code to be explained by ChatGPT. If not provided, it will use what's copied on
the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Rephrase text
}
\examples{
\dontrun{
cat(explain_code("for (i in 1:10) {\n  print(i ** 2)\n}"))
}

}
